refactor : step1 패키지 분리
