+++ src/main/java/racingcar/step1/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.step1;
 
 import java.util.Random;
