+++ src/test/java/racingcar/step1/CarTest.java
@@ -1,7 +1,8 @@
-package racingcar;
+package racingcar.step1;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.step1.Car;
 
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
