+++ src/test/java/racingcar/step1/RacingTest.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.step1;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.step1.Race;
 
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
